#1190 submissionDetails component
